manage -> utils; publish command reorg
