+++ packages/cli/src/commands/agent/utils/display.ts
