+++ packages/cli/src/commands/agent/utils/index.ts
