+++ packages/cli/src/commands/agent/utils/validation.ts
