+++ packages/cli/src/commands/create/utils/selection.ts
