+++ packages/cli/src/commands/create/utils/validation.ts
